refactor: cmd structure
